Refactoring of the refinement proposition
There is now a separate `REL e1 << e2 @ E : A` proposition,
on top of which we define the semantic typing judgement.

All the tactics work on the `REL .. << ..` proposition, and we use it
to prove the fundamental property. This means that we have to deal
with substitutions whenever we are doing something with the typing
judgement, but in practice this is not so bad.

Also added and updated: the counters refinement example.